refactor(card): extract durable writer calls
